# type-contract: remove sc-cache (was: fix empty cache typo)