--- a/typed-racket-test/unit-tests/contract-tests.rkt
+++ b/typed-racket-test/unit-tests/contract-tests.rkt
@@ -891,22 +891,4 @@
    (t (-prefab-top 'point 2))
    (t (-prefab-top '(box-box #(0)) 1))
 
-   (test-case "sc-cache hit"
-     (define sc-cache (make-hash))
-     (void
-       (type->contract -Number tc-fail #:typed-side #t #:sc-cache sc-cache))
-     (check-false (hash-empty? sc-cache))
-     (check-not-false (hash-ref sc-cache (cons -Number 'typed) #f)))
-
-   (test-case "sc-cache miss"
-     (define sc-cache (make-hash))
-     (define t
-       ;; cache the full recursive type,
-       ;;  but not any type that depends on the bound variable
-       (-mu X (Un -Null -String (-pair -Symbol X))))
-     (void
-       (type->contract t tc-fail #:typed-side #t #:sc-cache sc-cache))
-     (check-false (hash-empty? sc-cache))
-     (check-not-false (hash-ref sc-cache (cons t 'typed) #f))
-     (check-false (hash-ref sc-cache (cons (-pair -Symbol (-v X)) 'typed) #f)))
    ))