the old test (not (null? fv)) was always #t because fv is a function,
so `cached-match` never added anything to the cache

- - - 

two questions:

1. any idea how to test this? EDIT: 05dda60 has a test
2. is there a simpler function than `fv` that works here? I noticed `fv` builds a set then converts that to a list ... maybe `(equal? empty-free-vars (free-vars* type))` is better? EDIT: see below, `fv` isn't enough