@samth 2020-03-17
I'm ok with lack of testing, but I want to be sure that this is an optimization. Do you have any benchmark results?