@bennn 2020-03-17
I'll run the `tr-performance` benchmarks & post here.

But first, the test that inspired the `fv` check is failing (`..../succeed/pr15144.rkt`). It looks like a closed type can make an sc with free variables.