@rfindler 2020-03-18
Could you expose and query the cache?