@samth 2020-03-18
This looks good modulo performance checking.