@bennn 2020-03-18
Well, looks like this isn't an optimization.

- the `master1` and `master2` columns show the average of 5 compile-times on TR master
- `tc-cache1` and `tc-cache2` show average compile times on this branch
- all times are about the same; `tc-cache` is a little better for `schml-interp-casts-help` and worse for `dungeon`

So it looks like we should remove this cache

  benchmark               | master1         | master2         | tc-cache1      | tc-cache2       
  ------------------------|-----------------|-----------------|----------------|-----------------
  schml-specify-rep       | 7.79s (σ 0.07)  | 7.74s (σ 0.08)  | 7.77s (σ 0.05) | 7.81s (σ 0.03)  
  schml-interp-casts-help | 18.25s (σ 0.13) | 18.09s (σ 0.12) | 17.94s (σ 0.1) | 17.76s (σ 0.15) 
  parser                  | 1.97s (σ 0.01)  | 2.0s (σ 0.04)   | 1.98s (σ 0.01) | 1.98s (σ 0.01)  
  old-metrics             | 2.35s (σ 0.02)  | 2.38s (σ 0.04)  | 2.36s (σ 0.02) | 2.35s (σ 0.01)  
  new-metrics             | 3.18s (σ 0.04)  | 3.19s (σ 0.03)  | 3.15s (σ 0.02) | 3.15s (σ 0.02)  
  math-flonum             | 3.45s (σ 0.03)  | 3.45s (σ 0.03)  | 3.44s (σ 0.01) | 3.43s (σ 0.03)  
  fsm                     | 4.11s (σ 0.01)  | 4.14s (σ 0.05)  | 4.16s (σ 0.07) | 4.1s (σ 0.01)   
  forth                   | 4.01s (σ 0.03)  | 4.05s (σ 0.05)  | 4.08s (σ 0.07) | 4.04s (σ 0.02)  
  dungeon                 | 7.19s (σ 0.1)   | 7.14s (σ 0.07)  | 7.22s (σ 0.04) | 7.24s (σ 0.1)   
  bernoulli               | 3.15s (σ 0.05)  | 3.13s (σ 0.04)  | 3.13s (σ 0.02) | 3.14s (σ 0.05)  
  acquire                 | 13.25s (σ 0.1)  | 13.24s (σ 0.11) | 13.17s (σ 0.1) | 13.32s (σ 0.06)