@samth 2020-03-19
Is it possible to write a program that would benefit from this? And can you test some of the very large GUI type files?