@mfelleisen 2020-03-19
Some of these GUI-type things might be mine. 

I am beginning to wonder whether we could use 7GUIs for that.