@bennn 2020-03-19
Ok, I can check times for:

- `raco setup plot`
- `raco make 7GUI/Typed/*rkt`
- `raco make` for the biggest files under `typed-racket-more` : 
```
 typed-racket-more/typed/racket/private/gui-types.rkt 131475
 typed-racket-more/typed/private/framework-types.rkt   74266
 typed-racket-more/typed/framework.rkt                 34747
 typed-racket-more/typed/racket/gui/base.rkt           15123
 typed-racket-more/typed/pict.rkt                      11539
 typed-racket-more/typed/web-server/http.rkt           10882
```
Let me know if you have other ideas.