@samth 2020-03-19
That's basically what I suggest. I would look at particular files in `plot` though.