@bennn 2020-03-19
More results. The cache still doesn't seem to help.

15 programs total:

- the 2 `plot` programs provide the main renderers;
  I changed the `unsafe-provide` in those files to a normal `provide`
- the 7 "tasks" are from 7GUI
- the final 6 programs are from `typed-racket-more`

The `master` and "commit" columns show the average and standard deviation (in
parentheses) of 11 timings. Each timing comes from deleting the compiled folder
in the current directory and running `time raco make` on the file.

master = no cache, "commit" = cache

There's a `W` in the last column if the cache is significantly faster and a
"L" if it's significantly slower.

Most averages are within one stddev of each other, but there are two small wins
and two small losses.

  benchmark           | master       | 05dda60      | W/L
  --------------------|--------------|--------------|----
  plot2d.rkt          | 10.00 (0.06) | 10.28 (0.06) | L  
  plot3d.rkt          | 10.48 (0.05) | 10.78 (0.08) | L  
  task-1.rkt          | 2.45 (0.03)  | 2.46 (0.03)  |    
  task-2.rkt          | 3.00 (0.02)  | 2.98 (0.02)  |    
  task-3.rkt          | 3.31 (0.04)  | 3.32 (0.05)  |    
  task-4.rkt          | 2.48 (0.02)  | 2.46 (0.01)  |    
  task-5.rkt          | 2.70 (0.02)  | 2.71 (0.02)  |    
  task-6.rkt          | 8.88 (0.06)  | 8.55 (0.05)  | W  
  task-7.rkt          | 9.68 (0.08)  | 9.32 (0.08)  | W  
  gui-types.rkt       | 3.45 (0.04)  | 3.44 (0.02)  |    
  framework-types.rkt | 6.94 (0.04)  | 6.98 (0.05)  |    
  framework.rkt       | 2.99 (0.02)  | 3.05 (0.09)  |    
  gui/base.rkt        | 2.07 (0.02)  | 2.09 (0.03)  |    
  pict.rkt            | 1.83 (0.02)  | 1.82 (0.01)  |    
  http.rkt            | 3.58 (0.04)  | 3.56 (0.03)  |    

My interpretation is that cache hits only help a little because this type-to-
static-contract fold is just one of many places where TR needs to manipulate
a large type.

The cache might do better if it stored less --- keep only the top-level calls,
instead of every inner recursive call to `type->static-contract` --- but I
don't have great faith.