@mfelleisen 2020-03-20
It's interesting that two pairs of similarly long computations produce each a losing and a winning result.