@bennn 2020-03-27
@samth related followup question: why doesn't `core.rkt` create one definitions cache to share across all 3 contract passes?

https://github.com/racket/typed-racket/blob/master/typed-racket-lib/typed-racket/core.rkt#L43-L48

I've played with this a little and it seems fine --- but I haven't run all tests.